+++ CHANGELOG.md
@@ -7,8 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.8.6] - 2023-08-26
 
-- ``Biglist.DEFAULT_STORAGE_FORMAT`` was changed from ``pickle-zstd`` to ``pickle-z`` because the former
-  is available only in optional dependencies. This does not affect any previously persisted Biglist objects.
+- Made ``zstandard`` a required (as opposed to optional) dependency, because ``Biglist.DEFAULT_STORAGE_FORMAT`` defaults to ``pickle-zstd``.
 - Decreased default value of ``Biglist._n_write_threads`` from 8 to 4.