+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [0.8.6] - 2023-08-26
+
+- Made ``zstandard`` a required (as opposed to optional) dependency, because ``Biglist.DEFAULT_STORAGE_FORMAT`` defaults to ``pickle-zstd``.
+- Decreased default value of ``Biglist._n_write_threads`` from 8 to 4.
+
+
 ## [0.8.5] - 2023-08-25
 
 - ``BiglistBase`` now has custom support for pickling.